Fix Windows SDK and VS version checks
If WindowsTargetPlatformVersion is not set, the Visual Studio 15 (2017)
toolchain assumes that Windows 8.1 is being targeted. Since ARM64
support is only present and unlocked in Windows SDKs >= Windows 10 1809,
set that SDK as required in the vcxproj files.

Note that this will not be an issue in Visual Studio 16 or greater,
hence the -eq major version check.

https://developercommunity.visualstudio.com/content/problem/128836/windowstargetplatformversion-to-use-the-latest-ava.html

Bug: chromium:893460

Change-Id: Ib069501ad384d91349b1f635722dedd31a4edd97