Change to direction of scan for GF only group boost.
When coding a GF only group it makes more sense to scan forward
from the GF to choose the boost level rather than backwards from
the end of the group towards the GF.

In practice we do not often code GF only groups in normal 2 pass
encodes and when we do the video is usually almost static which means
the direction does not matter much. However,  a forward scan makes
more sense and is how things used to work before we started using
arfs most of the time.

Change-Id: I64a5a731ff579c8af86d8a6718830d426b16a755