vp9-rtc: Change init level of denoiser & noise level
Change to init/reset level of the denoiser from
kDenLow to kDenMedium, and the init noise level to kLow.
This affects the denoiser level during the initialization
stage of the noise estimation.

Improves denoising for noisy content during init stage of
noise estimation, with little effect for low noise/clean content.

Change-Id: I247a17b0f01f646fc2e91a4a070ad69bdb788cae