Further accelerate Armv8.0 Neon 4-tap convolution
Refactor Armv8.0 Neon 4-tap convolution functions to operate on 8-bit
types directly, rather than first widening to 16-bit.

2-tap (bilinear) filter values are always positive, but 4-tap filter
values are negative on the outer edges (taps 0 and 3), with taps 1
and 2 having much greater positive values to compensate. To use
instructions that operate on 8-bit types we also need the types to be
unsigned. In the convolution kernel, subtracting the products of taps
0 and 3 from the products of taps 1 and 2 always works since 2-tap
filters are 0-padded.

Co-authored by: Hari Limaye <hari.limaye@arm.com>

Change-Id: I87b32e2ef8cbd21eebb8cd2642e8826b704905b1