Remove unneeded assert in vpx_filter.h
Filter tap reporting was made more granular recently[1] to enable Arm
Neon optimizations that specialise convolution implementations
according to the filter size. This patch removes an assert that
should have been removed during that change - it no longer serves any
purpose to assert that the filter being used is a no-op filter.

This change is a pre-requisite for some highbd Neon convolution
changes that specialise implementations according to filter size.
(Without this change a convolve-copy test would fail should we
interrogate the size of the filter.)

[1] https://chromium-review.googlesource.com/c/webm/libvpx/+/5063929

Change-Id: I2a71680d27134535e6c0663b1668ba1b150b1a6f